+++ service-map/src/main/java/fi/nls/oskari/map/analysis/service/AnalysisDbService.java
@@ -7,11 +7,11 @@ import fi.nls.oskari.service.ServiceException;
 import fi.nls.oskari.service.db.BaseService;
 
 public interface AnalysisDbService extends BaseService<Analysis> {    
-        
-       
+
         public long insertAnalysisRow(final Analysis analysis);
         public int updateAnalysisCols(final Analysis analysis);
         public Analysis getAnalysisById(long id);
+        public List<Analysis> getAnalysisById(List<Long> idList);
         public List<Analysis> getAnalysisByUid(String uid);
         public void deleteAnalysisById(final long id) throws ServiceException;
         public void deleteAnalysis(final Analysis analysis) throws ServiceException;